requirements.txt
numpy
matplotlib
ase
jupyter
quippy-ase